+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Added
+
+- Added mod sha256 digest verification
+
 ## [0.9.9] 2024-03-17
 
 ### Changed